feat: major tree rework
- Reworked whole scene system into tree with various nodes. The goal
was to make addition of new items more streamlined, but for now its
not a real scene graph yet, since position of items is not based on
position of their parent - working on that.
- Added many node types - for static images, animations, text.
- Update cycle of tree's items has been split into update() and draw().
Draw shouldnt be used directly - its a method, called by scene tree
itself, if node has been added to it's draw queue (which should always
happen at the end of update cycle, if node is visible).
- Added some basic 2D camera. For the time being, it doesnt affect
render - e.g items will be added to draw queue even if they appear out
of POV.
- Added basic parallax support (use "distance" init kwarg, which is
presented in majority of visible nodes, to achieve it).
- Some other improvements and reworks I've already forgot about, which
make this update incompatible with previous versions of library.